referral/requirements.txt
asgiref==3.7.2
Django==3.2.3
django-sms==0.6.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.3
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
